requirements.txt
Flask>=3.0.0
Werkzeug>=3.0.1
Pillow>=10.0.0
opencv-python>=4.8.0
requests>=2.31.0
numpy>=1.24.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
google-genai>=1.46.0
flask-login>=0.6.2
flask-sqlalchemy>=3.0.5
flask-mail>=0.9.1
flask-wtf>=1.1.1
email-validator>=2.0.0
